summarizer/requirements.txt
dagster
dagster-postgres
psycopg2-binary
requests
openai
python-dotenv
